refactor(cli): split sandbox.ts into backend modules [C-01]
